refactor: move CLI commands init and run to commands folder
